Change account setup create account feature to theme 2
